api/requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==2.3.6
